+++ sdk/python/pulumi_kubernetes/authentication/v1/outputs.py
@@ -88,12 +88,12 @@ class TokenRequestSpec(dict):
     def __init__(__self__, *,
                  audiences: Sequence[str],
                  bound_object_ref: Optional['outputs.BoundObjectReference'] = None,
-                 expiration_seconds: Optional[float] = None):
+                 expiration_seconds: Optional[int] = None):
         """
         TokenRequestSpec contains client provided parameters of a token request.
         :param Sequence[str] audiences: Audiences are the intendend audiences of the token. A recipient of a token must identitfy themself with an identifier in the list of audiences of the token, and otherwise should reject the token. A token issued for multiple audiences may be used to authenticate against any of the audiences listed but implies a high degree of trust between the target audiences.
         :param 'BoundObjectReferenceArgs' bound_object_ref: BoundObjectRef is a reference to an object that the token will be bound to. The token will only be valid for as long as the bound object exists. NOTE: The API server's TokenReview endpoint will validate the BoundObjectRef, but other audiences may not. Keep ExpirationSeconds small if you want prompt revocation.
-        :param float expiration_seconds: ExpirationSeconds is the requested duration of validity of the request. The token issuer may return a token with a different validity duration so a client needs to check the 'expiration' field in a response.
+        :param int expiration_seconds: ExpirationSeconds is the requested duration of validity of the request. The token issuer may return a token with a different validity duration so a client needs to check the 'expiration' field in a response.
         """
         pulumi.set(__self__, "audiences", audiences)
         if bound_object_ref is not None:
@@ -119,7 +119,7 @@ class TokenRequestSpec(dict):
 
     @property
     @pulumi.getter(name="expirationSeconds")
-    def expiration_seconds(self) -> Optional[float]:
+    def expiration_seconds(self) -> Optional[int]:
         """
         ExpirationSeconds is the requested duration of validity of the request. The token issuer may return a token with a different validity duration so a client needs to check the 'expiration' field in a response.
         """
